requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
scikit-learn==1.0.2
pandas==1.3.3
numpy==1.21.2
python-docx==0.8.11
joblib==1.1.0
jinja2==3.0.1 